// src/app/shared/helpers/interfaces/products.interface.ts
export interface Products {
  Description: string;
  HaveOnStock: boolean;
  Image: string;
  IsDeleted: boolean;
  ManufacturerID: string;
  Percentage: number;
  Price: number;
  ProductGroupID: string;
  ProductID: string
  ProductTypeID: string;
  Promo: boolean;
  Quantity: number;
  Title:string;
}